refactor(perl-regex): split crate into SRP modules (#0000)
